refactor: split helpers.ts into focused renderer modules and remove dead code
Break the 1,449-line helpers.ts god-file into 9 single-purpose renderers
under src/renderers/ and 2 shared utils under src/utils/. Drop unused
duplicate implementations (parsers/textParser.ts, renderer stubs,
bookmarkUtils duplicate, types.d.ts), the dead processLink function,
and the stale defaultStyle/headingConfigs exports in types.ts. Fix the
contradictory include/exclude in tsconfig.json, tighten
MarkdownConversionError.context to unknown, wire processImage through
computeImageDimensions instead of re-inlining the same math, and add
TODO.md tracking the remaining follow-ups (inline parser rewrite,
bookmark ID collisions, Style restructure, lint tooling, etc.).

Public API, behavior, and all 57 tests unchanged.

Made-with: Cursor
