+++ src/renderers/commentRenderer.ts
@@ -1,5 +1,6 @@
 import { Paragraph, TextRun } from "docx";
 import { Style } from "../types.js";
+import { resolveFontFamily } from "../utils/styleUtils.js";
 
 /**
  * Processes a comment and returns appropriate paragraph formatting
@@ -8,12 +9,14 @@ import { Style } from "../types.js";
  * @returns The processed paragraph
  */
 export function processComment(text: string, style: Style): Paragraph {
+  const fontFamily = resolveFontFamily(style);
   return new Paragraph({
     children: [
       new TextRun({
         text: "Comment: " + text,
         italics: true,
         color: "666666",
+        font: fontFamily,
       }),
     ],
     spacing: {
