+++ src/renderers/headingRenderer.ts
@@ -1,7 +1,208 @@
-import { Paragraph, HeadingLevel, Bookmark, AlignmentType } from "docx";
+import {
+  Paragraph,
+  TextRun,
+  HeadingLevel,
+  AlignmentType,
+  Bookmark,
+} from "docx";
 import { Style, HeadingConfig } from "../types.js";
 import { sanitizeForBookmarkId } from "../utils/bookmarkUtils.js";
-import { processFormattedTextForHeading } from "../parsers/textParser.js";
+import { resolveFontFamily } from "../utils/styleUtils.js";
+
+/**
+ * Processes formatted text specifically for headings (bold/italic/underline/strikethrough)
+ * and returns an array of TextRun objects
+ */
+function processFormattedTextForHeading(
+  text: string,
+  fontSize: number,
+  style?: Style
+): TextRun[] {
+  const textRuns: TextRun[] = [];
+  let currentText = "";
+  let isBold = false;
+  let isItalic = false;
+  let isUnderline = false;
+  let isStrikethrough = false;
+
+  let boldStart = -1;
+  let italicStart = -1;
+  let underlineStart = -1;
+  let strikethroughStart = -1;
+  const fontFamily = resolveFontFamily(style);
+
+  function hasUnescapedMarker(src: string, marker: string, startIndex: number): boolean {
+    const maxIndex = src.length - marker.length;
+    for (let i = startIndex; i <= maxIndex; i++) {
+      if (src[i] === "\\") {
+        i++;
+        continue;
+      }
+      if (src.substring(i, i + marker.length) === marker) {
+        return true;
+      }
+    }
+    return false;
+  }
+
+  function createRun(value: string): TextRun {
+    return new TextRun({
+      text: value,
+      bold: isBold,
+      italics: isItalic,
+      strike: isStrikethrough,
+      underline: isUnderline ? { type: "single" } : undefined,
+      color: "000000",
+      size: fontSize,
+      font: fontFamily,
+      rightToLeft: style?.direction === "RTL",
+    });
+  }
+
+  function flushCurrentText(): void {
+    if (!currentText) {
+      return;
+    }
+    textRuns.push(createRun(currentText));
+    currentText = "";
+  }
+
+  for (let j = 0; j < text.length; j++) {
+    if (text[j] === "\\" && j + 1 < text.length) {
+      const nextChar = text[j + 1];
+      if (nextChar === "*" || nextChar === "+" || nextChar === "~" || nextChar === "\\") {
+        currentText += nextChar;
+        j++;
+        continue;
+      }
+      currentText += text[j];
+      continue;
+    }
+
+    if (j + 2 < text.length && text[j] === "*" && text[j + 1] === "*" && text[j + 2] === "*") {
+      flushCurrentText();
+      if (!isBold && !isItalic) {
+        boldStart = j;
+        italicStart = j;
+      } else {
+        boldStart = -1;
+        italicStart = -1;
+      }
+      isBold = !isBold;
+      isItalic = !isItalic;
+      j += 2;
+      continue;
+    }
+
+    if (j + 1 < text.length && text[j] === "*" && text[j + 1] === "*") {
+      flushCurrentText();
+      if (!isBold) {
+        boldStart = j;
+      } else {
+        boldStart = -1;
+      }
+      isBold = !isBold;
+      j++;
+      continue;
+    }
+
+    if (j + 1 < text.length && text[j] === "+" && text[j + 1] === "+") {
+      const canToggle = isUnderline || hasUnescapedMarker(text, "++", j + 2);
+      if (!canToggle) {
+        currentText += "++";
+        j++;
+        continue;
+      }
+      flushCurrentText();
+      if (!isUnderline) {
+        underlineStart = j;
+      } else {
+        underlineStart = -1;
+      }
+      isUnderline = !isUnderline;
+      j++;
+      continue;
+    }
+
+    if (j + 1 < text.length && text[j] === "~" && text[j + 1] === "~") {
+      const canToggle = isStrikethrough || hasUnescapedMarker(text, "~~", j + 2);
+      if (!canToggle) {
+        currentText += "~~";
+        j++;
+        continue;
+      }
+      flushCurrentText();
+      if (!isStrikethrough) {
+        strikethroughStart = j;
+      } else {
+        strikethroughStart = -1;
+      }
+      isStrikethrough = !isStrikethrough;
+      j++;
+      continue;
+    }
+
+    if (
+      text[j] === "*" &&
+      (j === 0 || text[j - 1] !== "*") &&
+      (j === text.length - 1 || text[j + 1] !== "*")
+    ) {
+      flushCurrentText();
+      if (!isItalic) {
+        italicStart = j;
+      } else {
+        italicStart = -1;
+      }
+      isItalic = !isItalic;
+      continue;
+    }
+
+    currentText += text[j];
+  }
+
+  if (currentText) {
+    if (isBold && isItalic && boldStart >= 0 && italicStart >= 0 && boldStart === italicStart) {
+      currentText = "***" + currentText;
+      isBold = false;
+      isItalic = false;
+    } else {
+      if (isBold && boldStart >= 0) {
+        currentText = "**" + currentText;
+        isBold = false;
+      }
+      if (isItalic && italicStart >= 0) {
+        currentText = "*" + currentText;
+        isItalic = false;
+      }
+      if (isUnderline && underlineStart >= 0) {
+        currentText = "++" + currentText;
+        isUnderline = false;
+      }
+      if (isStrikethrough && strikethroughStart >= 0) {
+        currentText = "~~" + currentText;
+        isStrikethrough = false;
+      }
+    }
+
+    if (currentText.trim()) {
+      textRuns.push(createRun(currentText));
+    }
+  }
+
+  if (textRuns.length === 0) {
+    textRuns.push(
+      new TextRun({
+        text: "",
+        color: "000000",
+        size: fontSize,
+        bold: true,
+        font: fontFamily,
+      })
+    );
+  }
+
+  return textRuns;
+}
 
 /**
  * Processes a heading line and returns appropriate paragraph formatting and a bookmark ID
@@ -19,18 +220,18 @@ export function processHeading(
 ): { paragraph: Paragraph; bookmarkId: string } {
   const headingText = line.replace(new RegExp(`^#{${config.level}} `), "");
   const headingLevel = config.level;
-  // Generate a unique bookmark ID using the clean text (without markdown)
+
   const cleanTextForBookmark = headingText
     .replace(/\*\*/g, "")
-    .replace(/\*/g, "");
+    .replace(/\*/g, "")
+    .replace(/\+\+/g, "")
+    .replace(/~~/g, "");
   const bookmarkId = `_Toc_${sanitizeForBookmarkId(
     cleanTextForBookmark
   )}_${Date.now()}`;
 
-  // Get the appropriate font size based on heading level and custom style
   let headingSize = style.titleSize;
 
-  // Use specific heading size if provided, otherwise calculate based on level
   if (headingLevel === 1 && style.heading1Size) {
     headingSize = style.heading1Size;
   } else if (headingLevel === 2 && style.heading2Size) {
@@ -42,37 +243,30 @@ export function processHeading(
   } else if (headingLevel === 5 && style.heading5Size) {
     headingSize = style.heading5Size;
   } else if (headingLevel > 1) {
-    // Fallback calculation if specific size not provided
     headingSize = style.titleSize - (headingLevel - 1) * 4;
   }
 
-  // Determine alignment based on heading level
   let alignment;
-
-  // Check for level-specific alignment first
   if (headingLevel === 1 && style.heading1Alignment) {
-    alignment = AlignmentType[style.heading1Alignment as keyof typeof AlignmentType];
+    alignment = AlignmentType[style.heading1Alignment];
   } else if (headingLevel === 2 && style.heading2Alignment) {
-    alignment = AlignmentType[style.heading2Alignment as keyof typeof AlignmentType];
+    alignment = AlignmentType[style.heading2Alignment];
   } else if (headingLevel === 3 && style.heading3Alignment) {
-    alignment = AlignmentType[style.heading3Alignment as keyof typeof AlignmentType];
+    alignment = AlignmentType[style.heading3Alignment];
   } else if (headingLevel === 4 && style.heading4Alignment) {
-    alignment = AlignmentType[style.heading4Alignment as keyof typeof AlignmentType];
+    alignment = AlignmentType[style.heading4Alignment];
   } else if (headingLevel === 5 && style.heading5Alignment) {
-    alignment = AlignmentType[style.heading5Alignment as keyof typeof AlignmentType];
+    alignment = AlignmentType[style.heading5Alignment];
   } else if (style.headingAlignment) {
-    // Fallback to general heading alignment if no level-specific alignment
-    alignment = AlignmentType[style.headingAlignment as keyof typeof AlignmentType];
+    alignment = AlignmentType[style.headingAlignment];
   }
 
-  // Process the heading text to handle markdown formatting (bold/italic)
   const processedTextRuns = processFormattedTextForHeading(
     headingText,
     headingSize,
     style
   );
 
-  // Create the paragraph with bookmark
   const paragraph = new Paragraph({
     children: [
       new Bookmark({
@@ -88,7 +282,7 @@ export function processHeading(
       after: style.headingSpacing / 2,
     },
     alignment: alignment,
-    style: `Heading${headingLevel}`, // This is crucial for TOC recognition
+    style: `Heading${headingLevel}`,
     bidirectional: style.direction === "RTL",
   });
