--- src/types.d.ts
@@ -1,8 +1,0 @@
-declare global {
-  interface Window {
-    URL: {
-      createObjectURL(blob: Blob): string;
-      revokeObjectURL(url: string): void;
-    };
-  }
-}
